Set transparent background for menu-items-section in media menu
The menu-items-section box of the media menu was not transparent by
default. Usually it is fully shadowed by its child "menu-content".
The color and tint of the can be configured by the user.
However, because the menu-items-section box behind the card had a
solid color, it was impossible to configure a transparent card.