Ensure notification service is available when sending a notification (#720)
* Ensure notification service is available when sending a notification

This is a bit of a weird patch for a fundamental order-of-execution
issue.

The bundler (esbuild) can sometimes place the `notifdService`
instantiation bellow the code calling `Notify`, which would causes the
`notify-send` execution to fail. To avoid this, we can make it seem like
`notifdService` is being used inside `Notify`, so that it is always
placed above the usage of the function, ensuring that we always have a
notification daemon running

* Add eslint-disable-line

* Update src/lib/utils.ts

---------

Co-authored-by: Jas Singh <jaskiratpal.singh@outlook.com>